Workspace.updateProjectMetaData is buggy because it creates a new Project instance while older instances may still exist. Moved method to the Project class.
